detect-tls-ja3-string: move unittests to tests/
